Clean up spacing. Move Header to views folder. Create Footer component.
