+++ src/views/Footer.vue
@@ -1,0 +1,13 @@
+<template>
+    <footer>
+        <nav class='footer-links'>
+            <router-link class='footer-link' to='/about'>About</router-link>
+        </nav>
+    </footer>
+</template>
+
+<script>
+export default {
+    name: 'Footer'
+}
+</script>
