+++ src/views/Header.vue
@@ -19,7 +19,8 @@
 </template>
 
 <script>
-import router from './router';
+import router from '../router';
+
 export default {
   name: "Header",
   props: ['onHome' , 'todaysDate'],
