+++ src/views/PicOfTheDay.vue
@@ -36,7 +36,7 @@
 </template>
 
 <script>
-import Preview from './Preview.vue'
+import Preview from './Preview.vue';
 import moment from 'moment';
 
 export default {
